[cw|#22] add support for src and resource files
Major Changes:
* introduce public `Code` struct
* modify `(*context).Evaluate()` & `(*sandbox).run()` to accept `Code`
  struct
* introduce logic to move source files and resource files into tmp
  mounted docker directory
* refactored `cmd` and tests

Notes:
since we will be allowing the user's program to potentially write to a
resource file, we should add some security to ensure that the program
isn't dumping data into these files in an unbounded way. a malicious
program might want to do that to use up al storage.